% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{criteria}
\alias{criteria}
\alias{is_rstudio_project}
\alias{is_vscode_project}
\alias{is_r_package}
\alias{is_remake_project}
\alias{is_drake_project}
\alias{is_targets_project}
\alias{is_pkgdown_project}
\alias{is_renv_project}
\alias{is_projectile_project}
\alias{is_quarto_project}
\alias{is_git_root}
\alias{is_svn_root}
\alias{is_vcs_root}
\alias{is_testthat}
\alias{from_wd}
\title{Prespecified criteria}
\usage{
is_rstudio_project

is_vscode_project

is_r_package

is_remake_project

is_drake_project

is_targets_project

is_pkgdown_project

is_renv_project

is_projectile_project

is_quarto_project

is_git_root

is_svn_root

is_vcs_root

is_testthat

from_wd

criteria
}
\description{
This is a collection of commonly used root criteria.
}
\details{
\code{is_rstudio_project} looks for a file with extension \code{.Rproj}.

\code{is_vscode_project} looks for a \code{.vscode/settings.json} file.

\code{is_r_package} looks for a \code{DESCRIPTION} file.

\code{is_remake_project} looks for a \code{remake.yml} file.

\code{is_drake_project} looks for a \code{.drake} directory.

\code{is_targets_project} looks for a \verb{_targets.R} file.

\code{is_pkgdown_project} looks for a \verb{_pkgdown.yml}, \verb{_pkgdown.yaml}, \verb{pkgdown/_pkgdown.yml} and/or \verb{inst/_pkgdown.yml} file.

\code{is_renv_project} looks for an \code{renv.lock} file.

\code{is_projectile_project} looks for a \code{.projectile} file.

\code{is_quarto_project} looks for a \verb{_quarto.yml} file.

\code{is_git_root} looks for a \code{.git} directory.

\code{is_svn_root} looks for a \code{.svn} directory.

\code{is_vcs_root} looks for the root of a version control system,
currently only Git and SVN are supported.

\code{is_testthat} looks for the \code{testthat} directory, works when developing, testing, and checking a package.

\code{from_wd} uses the current working directory.
}
