% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{as.root_criterion}
\alias{is.root_criterion}
\title{Deprecated functions}
\usage{
as.root_criterion(...)

is.root_criterion(...)
}
\description{
Use \code{\link[=as_root_criterion]{as_root_criterion()}} and \code{\link[=is_root_criterion]{is_root_criterion()}}, respectively.
}
\keyword{internal}
