% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-file.R
\docType{data}
\name{criteria}
\alias{criteria}
\alias{from_wd}
\alias{is_r_package}
\alias{is_remake_project}
\alias{is_rstudio_project}
\alias{is_testthat}
\title{Prespecified criteria}
\format{An object of class \code{root_criteria} of length 5.}
\usage{
criteria

is_rstudio_project

is_r_package

is_remake_project

is_testthat

from_wd
}
\description{
This is a collection of commonly used root criteria.
}
\details{
\code{is_rstudio_project} looks for a file with extension \code{.Rproj}.

\code{is_r_package} looks for a \code{DESCRIPTION} file.

\code{is_remake_project} looks for a \code{remake.yml} file.

\code{is_testthat} looks for the \code{testthat} directory, works when
  developing, testing, and checking a package.

\code{from_wd} uses the current working directory.
}
\keyword{datasets}

