% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/has-file.R
\docType{data}
\name{criteria}
\alias{criteria}
\alias{is_r_package}
\alias{is_rstudio_project}
\title{Prespecified criteria}
\format{\preformatted{List of 2
 $ is_rstudio_project: chr "Root criterion: Contains a file matching '^.*\\\\.Rproj$' with contents matching '^Version: ' in the first 1 lines"
 $ is_r_package      : chr "Root criterion: Contains a file 'DESCRIPTION' with contents matching '^Package: '"
}}
\usage{
criteria

is_rstudio_project

is_r_package
}
\description{
This is a collection of commonly used root criteria.
}
\details{
\code{is_rstudio_project} looks for a file with extension \code{.Rproj}.

\code{is_r_package} looks for a \code{DESCRIPTION} file.
}
\keyword{datasets}

