% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R, R/deprecated.R
\name{is.root_criterion}
\alias{is.root_criterion}
\alias{as.root_criterion}
\alias{deprecated}
\title{Deprecated functions}
\usage{
is.root_criterion(x)

as.root_criterion(x)
}
\arguments{
\item{x}{An object}
}
\description{
See the "Functions" section for replacements
}
\section{Functions}{

\itemize{
\item \code{as.root_criterion()}: Use \code{\link[=as_root_criterion]{as_root_criterion()}}  instead.
\item \code{is.root_criterion()}: Use \code{\link[=is_root_criterion]{is_root_criterion()}} instead.
}
}

