% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/root.R
\name{find_root}
\alias{find_root}
\title{Find the root of a directory hierarchy}
\usage{
find_root(filename, contents = NULL, n = -1L, path = getwd(), ...)
}
\arguments{
\item{filename}{A regular expression to match the file name}

\item{contents}{A regular expression to match the file contents}

\item{n}{integer.  The (maximal) number of lines to
    read.  Negative values indicate that one should read up to the end of
    input on the connection.}

\item{path}{The start directory}

\item{...}{Additional params passed to \code{\link{file.path}}}
}
\value{
The normalized path of the root as specified by the search criteria.
  Throws an error if no root is found
}
\description{
Starting from the working directory, this function searches for a file
whose name matches a given pattern and which optionally contains a given text.
If such a file is not found, the search proceeds up the directory hierarchy.
}
\seealso{
\code{\link[utils]{glob2rx}} \code{\link{file.path}}
}

