% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/root.R
\name{find_root}
\alias{find_root}
\title{Find the root of a directory hierarchy}
\usage{
find_root(criterion, path = getwd())
}
\arguments{
\item{criterion}{A criterion, will be coerced using
\code{\link{as.root_criterion}}}

\item{path}{The start directory}
}
\value{
The normalized path of the root as specified by the search criterion.
  Throws an error if no root is found
}
\description{
A \emph{root} is defined as a directory that contains a regular file
whose name matches a given pattern and which optionally contains a given text.
The search for a root starts at a given directory (the working directory
by default), and proceeds up the directory hierarchy.
}
\details{
Starting from the working directory, the \code{find_root} function searches
for the root.
If a root is found, the \code{...} arguments are used to construct a path;
thus, if no extra arguments are given, the root is returned.
If no root is found, an error is thrown.
}
\examples{
\dontrun{
find_root(glob2rx("DESCRIPTION"), "^Package: ")
}
}
\seealso{
\code{\link[utils]{glob2rx}} \code{\link{file.path}}
}

