% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R, R/package-root.R, R/shortcut.R
\name{find_root_file}
\alias{find_root_file}
\alias{r_package_subdirs}
\alias{find_package_root_file}
\alias{find_rstudio_root_file}
\alias{find_remake_root_file}
\alias{find_testthat_root_file}
\title{File paths relative to the root of a directory hierarchy}
\usage{
find_root_file(..., criterion, path = ".")

r_package_subdirs()

find_package_root_file(..., path = ".", subdirs = r_package_subdirs())

find_rstudio_root_file(..., path = ".")

find_remake_root_file(..., path = ".")

find_testthat_root_file(..., path = ".")
}
\arguments{
\item{...}{\verb{[character]}\cr
Further path components passed to \code{\link[=file.path]{file.path()}}.
All arguments must be the same length or length one.}

\item{criterion}{\verb{[root_criterion]}\cr
A criterion, one of the predefined \link{criteria}
or created by \code{\link[=root_criterion]{root_criterion()}}.
Will be coerced using \code{\link[=as_root_criterion]{as_root_criterion()}}.}

\item{path}{\verb{[character(1)]}\cr
The start directory.}

\item{subdirs}{\verb{[character()]}\cr
Well-known subdirectories that may contain the R package,
consulted in the order given,
by default the value of \code{r_package_subdirs()}.
Pass \code{character()} to consult only \code{.Rproj} files,
or \code{NULL} to disable both and require a \code{DESCRIPTION} file in the
directories themselves.}
}
\value{
The normalized path of the root as specified by the search criteria,
with the additional path components appended.
Throws an error if no root is found.
}
\description{
\code{find_root_file()} is a wrapper around \code{\link[=find_root]{find_root()}} that
appends an arbitrary number of path components to the root using
\code{\link[base:file.path]{base::file.path()}}.
}
\details{
This function operates on the notion of relative paths.
The \code{...} argument is expected to contain a path relative to the root.
If the first path component passed to \code{...} is already an absolute path,
the \code{criterion} and \code{path} arguments are ignored,
and \code{...} is forwarded to \code{\link[=file.path]{file.path()}}.

\code{r_package_subdirs()} returns the well-known subdirectories that
\code{\link[=find_package_root_file]{find_package_root_file()}} consults when a directory in the hierarchy
contains no \code{DESCRIPTION} file itself.
The convention is shared with the \pkg{pkgdepends} package, which applies it
when installing from a Git remote.

\code{find_package_root_file()} also finds packages that live in a subdirectory
of a multi-language repository, e.g. a repository that contains an R package,
a Python package and a JavaScript library side by side.
For each directory in the hierarchy, the following are consulted in order:
\enumerate{
\item the directory itself, if it contains a \code{DESCRIPTION} file;
\item the directory given by the \verb{PackagePath:} field of an RStudio project
file (\code{.Rproj}) in that directory, if that directory contains a
\code{DESCRIPTION} file;
\item the \code{subdirs}, in the order given, the first one that contains a
\code{DESCRIPTION} file.
}

Pass \code{subdirs = NULL} to consult only the directories themselves, the
behavior of rprojroot 2.1.1 and earlier.
}
\examples{
\dontrun{
find_package_root_file("tests", "testthat.R")
has_file("DESCRIPTION", "^Package: ")$find_file
has_file("DESCRIPTION", "^Package: ")$make_fix_file(".")
}

r_package_subdirs()
}
\seealso{
\code{\link[=find_root]{find_root()}} \code{\link[utils:glob2rx]{utils::glob2rx()}} \code{\link[base:file.path]{base::file.path()}}
}
