% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/criterion.R, R/has-file.R
\name{root_criterion}
\alias{has_file}
\alias{root_criterion}
\title{Is a directory the project root?}
\usage{
root_criterion(testfun, desc)

has_file(filename, contents = NULL, n = -1L)
}
\arguments{
\item{testfun}{A function with one parameter that returns \code{TRUE}
if the directory specified by this parameter is the project root,
and \code{FALSE} otherwise}

\item{desc}{A textual description of the test criterion}

\item{filename,contents}{Regular expressions to match the file name or contents}

\item{n}{integer.  The (maximal) number of lines to
    read.  Negative values indicate that one should read up to the end of
    input on the connection.}
}
\description{
TBD
}

