% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprojroot-package.R
\docType{package}
\name{rprojroot-package}
\alias{rprojroot}
\alias{rprojroot-package}
\title{Finding Files in Project Subdirectories}
\description{
Robust, reliable and flexible paths to files below a
project root. The 'root' of a project is defined as a directory
that matches a certain criterion, e.g., it contains a certain
regular file.
}
\details{
See the "Value" section in \code{\link{root_criterion}} for documentation
of root criterion objects, and  \code{\link{criteria}} for useful predefined
root criteria.
}
\examples{
criteria
is_r_package$find_file("NAMESPACE")
root_fun <- is_r_package$make_fix_file()
root_fun("NAMESPACE")
}

