% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprojroot-package.R
\docType{package}
\name{rprojroot-package}
\alias{rprojroot}
\alias{rprojroot-package}
\title{rprojroot: Finding Files in Project Subdirectories}
\description{
Robust, reliable and flexible paths to files below a
project root. The 'root' of a project is defined as a directory
that matches a certain criterion, e.g., it contains a certain
regular file.
}
\details{
See the "Value" section in \code{\link[=root_criterion]{root_criterion()}} for documentation
of root criterion objects, and  \code{\link[=criteria]{criteria()}} for useful predefined
root criteria.
}
\examples{
criteria
\dontrun{
is_r_package$find_file("NAMESPACE")
root_fun <- is_r_package$make_fix_file()
root_fun("NAMESPACE")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/rprojroot}
  \item \url{https://r-lib.github.io/rprojroot}
  \item Report bugs at \url{https://github.com/r-lib/rprojroot/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (0000-0002-1416-3412)

}
